LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/credentials.sub-expected.txt



FAIL Modules should be loaded with or without the credentials based on the same-origin-ness and the crossOrigin attribute assert_equals: Modules should be loaded with the credentials when the crossOrigin attribute is specified with "use-credentials" as its value and the target is cross-origin expected "found" but got "not found"
